Remove redundant casts to float when retrieving model frames

+++ dlls/bullsquid.cpp
@@ -101,7 +101,7 @@ void CSquidSpit::SpawnHelper(const char *className)
 
 	UTIL_SetSize( pev, Vector( 0, 0, 0 ), Vector( 0, 0, 0 ) );
 
-	m_maxFrame = (float)MODEL_FRAMES( pev->modelindex ) - 1;
+	m_maxFrame = MODEL_FRAMES( pev->modelindex ) - 1;
 }
 
 void CSquidSpit::Animate( void )
@@ -237,7 +237,7 @@ void CBigSquidSpit::Spawn( void )
 
 	UTIL_SetSize( pev, Vector( 0, 0, 0 ), Vector( 0, 0, 0 ) );
 
-	m_maxFrame = (float)MODEL_FRAMES( pev->modelindex ) - 1;
+	m_maxFrame = MODEL_FRAMES( pev->modelindex ) - 1;
 }
 
 void CBigSquidSpit::Precache()
